[test] Enable failing unexpected console warn|error in browser tests
